Move StreamingBuffer to TableInfo, remove streamingBuffer from ExternalTableInfo
