+++ gcloud-java-bigquery/src/test/java/com/google/gcloud/bigquery/TableInfoTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import com.google.common.collect.ImmutableList;
-import com.google.gcloud.bigquery.BaseTableInfo.StreamingBuffer;
+import com.google.gcloud.bigquery.TableInfo.StreamingBuffer;
 
 import org.junit.Test;
 
@@ -46,7 +46,6 @@ public class TableInfoTest {
   private static final Schema TABLE_SCHEMA = Schema.of(FIELD_SCHEMA1, FIELD_SCHEMA2, FIELD_SCHEMA3);
   private static final String VIEW_QUERY = "VIEW QUERY";
   private static final List<String> SOURCE_URIS = ImmutableList.of("uri1", "uri2");
-  private static final String SOURCE_FORMAT = "CSV";
   private static final Integer MAX_BAD_RECORDS = 42;
   private static final Boolean IGNORE_UNKNOWN_VALUES = true;
   private static final String COMPRESSION = "GZIP";
@@ -96,7 +95,6 @@ public class TableInfoTest {
           .numBytes(NUM_BYTES)
           .numRows(NUM_ROWS)
           .selfLink(SELF_LINK)
-          .streamingBuffer(STREAMING_BUFFER)
           .build();
   private static final List<UserDefinedFunction> USER_DEFINED_FUNCTIONS =
       ImmutableList.of(UserDefinedFunction.inline("Function"), UserDefinedFunction.fromUri("URI"));
@@ -184,7 +182,6 @@ public class TableInfoTest {
     assertEquals(NUM_BYTES, EXTERNAL_TABLE_INFO.numBytes());
     assertEquals(NUM_ROWS, EXTERNAL_TABLE_INFO.numRows());
     assertEquals(SELF_LINK, EXTERNAL_TABLE_INFO.selfLink());
-    assertEquals(STREAMING_BUFFER, EXTERNAL_TABLE_INFO.streamingBuffer());
     assertEquals(BaseTableInfo.Type.EXTERNAL, EXTERNAL_TABLE_INFO.type());
   }
 
@@ -235,6 +232,5 @@ public class TableInfoTest {
     compareBaseTableInfo(expected, value);
     assertEquals(expected, value);
     assertEquals(expected.configuration(), value.configuration());
-    assertEquals(expected.streamingBuffer(), value.streamingBuffer());
   }
 }
